Move things around so veneur/trace doesn't pull in veneur/samplers
The samplers package needs a *ton* of other packages (it effectively
pulls in most of veneur), which makes depending on package trace
cumbersome. This change reduces the dep footprint to only trace &
protocol.
