+++ protocol/wire.go
@@ -29,13 +29,13 @@
 package protocol
 
 import (
+	"fmt"
 	"io"
 	"sync"
 
 	"encoding/binary"
 
 	"github.com/gogo/protobuf/proto"
-	"github.com/stripe/veneur/samplers"
 	"github.com/stripe/veneur/ssf"
 )
 
@@ -67,6 +67,55 @@ func readFrame(in io.Reader, length int) ([]byte, error) {
 	}
 }
 
+// A Message struct represents a parsed SSF message. It encapsulates
+// an ssf.SSFSpan object, which can contain a trace span and/or a set
+// of metrics.
+type Message struct {
+	span *ssf.SSFSpan
+}
+
+// InvalidTrace is an error type indicating that an SSF span was
+// invalid.
+type InvalidTrace struct {
+	span *ssf.SSFSpan
+}
+
+func (e *InvalidTrace) Error() string {
+	return fmt.Sprintf("not a valid trace span: %#v", e.span)
+}
+
+// TraceSpan checks if an SSF message is a valid trace. If so, it
+// returns a pointer to that original span. If the span is not a valid
+// trace, TraceSpan returns nil and an *InvalidTrace error type.
+//
+// The span returned from TraceSpan does contain the (unparsed) metric
+// samples contained in the span. Note also that since the data
+// returned is a pointer to the original span, it's not advisable to
+// modify that span directly.
+func (m *Message) TraceSpan() (*ssf.SSFSpan, error) {
+	if !ValidTrace(m.span) {
+		return nil, &InvalidTrace{m.span}
+	}
+	return m.span, nil
+}
+
+// ValidTrace takes in an SSF span and determines if it is valid or not.
+// It also makes sure the Tags is non-nil, since we use it later.
+func ValidTrace(sample *ssf.SSFSpan) bool {
+	ret := true
+	ret = ret && sample.Id != 0
+	ret = ret && sample.TraceId != 0
+	ret = ret && sample.StartTimestamp != 0
+	ret = ret && sample.EndTimestamp != 0
+	return ret
+}
+
+// Metrics returns the ssf samples contained in an SSF span. It does
+// not perform any parsing or conversion.
+func (m *Message) Metrics() []*ssf.SSFSample {
+	return m.span.Metrics
+}
+
 // ReadSSF reads a framed SSF span from a stream and returns a parsed
 // SSFSpan structure and a set of statsd metrics.
 //
@@ -75,8 +124,7 @@ func readFrame(in io.Reader, length int) ([]byte, error) {
 // unrecoverably broken. The error is EOF only if no bytes were read
 // at the start of a message (e.g. if a connection was closed after
 // the last message).
-
-func ReadSSF(in io.Reader) (*samplers.Message, error) {
+func ReadSSF(in io.Reader) (*Message, error) {
 	var version uint8
 	var length uint32
 	if err := binary.Read(in, binary.BigEndian, &version); err != nil {
@@ -100,7 +148,47 @@ func ReadSSF(in io.Reader) (*samplers.Message, error) {
 	if err != nil {
 		return nil, &errFramingIO{err}
 	}
-	return samplers.ParseSSF(bts)
+	return ParseSSF(bts)
+}
+
+// ParseSSF takes in a byte slice and returns: a normalized SSFSpan
+// and an error if any errors in parsing the SSF packet occur.
+func ParseSSF(packet []byte) (*Message, error) {
+	span := &ssf.SSFSpan{}
+	scratchBuff := pbufPool.Get().(*proto.Buffer)
+	defer func() {
+		scratchBuff.Reset()
+		pbufPool.Put(scratchBuff)
+	}()
+	scratchBuff.SetBuf(packet)
+	err := scratchBuff.Unmarshal(span)
+
+	if err != nil {
+		return nil, err
+	}
+
+	// Normalize the span:
+	if span.Tags == nil {
+		span.Tags = map[string]string{}
+	}
+	if span.Name == "" {
+		// Even though incoming packets should have Name set,
+		// this allows Veneur to be backwards-compatible.
+		for k, v := range span.Tags {
+			if k == "name" {
+				span.Name = v
+			}
+		}
+		delete(span.Tags, "name")
+	}
+
+	// Normalize metrics on the span:
+	for _, sample := range span.Metrics {
+		if sample.SampleRate == 0 {
+			sample.SampleRate = 1
+		}
+	}
+	return &Message{span}, nil
 }
 
 var pbufPool = sync.Pool{
