+++ trace/client_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stripe/veneur/protocol"
-	"github.com/stripe/veneur/samplers"
 )
 
 func TestNoClient(t *testing.T) {
@@ -60,7 +59,7 @@ func TestUNIX(t *testing.T) {
 	laddr, err := net.ResolveUnixAddr("unix", sockName)
 	require.NoError(t, err)
 
-	outPkg := make(chan *samplers.Message, 4)
+	outPkg := make(chan *protocol.Message, 4)
 	cleanup := serveUNIX(t, laddr, func(in net.Conn) {
 		for {
 			pkg, err := protocol.ReadSSF(in)
@@ -100,7 +99,7 @@ func TestUNIXBuffered(t *testing.T) {
 	laddr, err := net.ResolveUnixAddr("unix", sockName)
 	require.NoError(t, err)
 
-	outPkg := make(chan *samplers.Message, 4)
+	outPkg := make(chan *protocol.Message, 4)
 	cleanup := serveUNIX(t, laddr, func(in net.Conn) {
 		for {
 			pkg, err := protocol.ReadSSF(in)
@@ -204,7 +203,7 @@ func TestReconnectUNIX(t *testing.T) {
 	laddr, err := net.ResolveUnixAddr("unix", sockName)
 	require.NoError(t, err)
 
-	outPkg := make(chan *samplers.Message, 4)
+	outPkg := make(chan *protocol.Message, 4)
 	// A server that can read one span and then immediately closes
 	// the connection:
 	cleanup := serveUNIX(t, laddr, func(in net.Conn) {
@@ -272,7 +271,7 @@ func TestReconnectBufferedUNIX(t *testing.T) {
 	laddr, err := net.ResolveUnixAddr("unix", sockName)
 	require.NoError(t, err)
 
-	outPkg := make(chan *samplers.Message, 4)
+	outPkg := make(chan *protocol.Message, 4)
 	// A server that can read one span and then immediately closes
 	// the connection:
 	cleanup := serveUNIX(t, laddr, func(in net.Conn) {
